Transition to a Header-Only Library for ease of use (no setup process).
